ファイル整理、押下時ずっと鳴らすボタン作成
